pyproject.toml
[tool.poetry]
name = "pil-utils"
version = "0.2.0b4"
description = "A simple PIL wrapper and text-to-image tool"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/MeetWq/pil-utils"
repository = "https://github.com/MeetWq/pil-utils"

[tool.poetry.dependencies]
python = "^3.9"
Pillow = ">=10.0.0,<12.0.0"
numpy = ">=1.21.0"
opencv-python-headless = "^4.0.0"
bbcode = "^1.1.0"
skia-python = ">=130.0b10"

[tool.poetry.group.dev.dependencies]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
